Refactored the dashboard stylesheets directory
