+++ frontend/src/functionality/Dashboard/TaskSection/TaskOptionsDropup/TaskOptionsButton.jsx
@@ -1,6 +1,6 @@
 import useDropdown from '../../../../utils/Hooks/useDropdown.js';
 import TaskOptionsDropup from './TaskOptionsDropup.jsx';
-import '../../stylesheets/TaskOptionsButton.css';
+import '../../stylesheets/TaskSection/TaskOptionsDropup/TaskOptionsButton.css';
 
 /**
  * A button that toggles a drop-up menu of task options.
