+++ frontend/src/functionality/Dashboard/stylesheets/NotesSection/NotesSection.css
