+++ frontend/src/functionality/Dashboard/stylesheets/NotesSection/NotesTextArea.css
