+++ frontend/src/functionality/Dashboard/stylesheets/TaskSection/TaskItemContent.css
